# [PoC] Experiment: replace msgpack with protobuf for analytics