--- a/gateway/analytics_test.go
+++ b/gateway/analytics_test.go
@@ -1,9 +1,16 @@
 package gateway
 
 import (
+	"net/http"
 	"testing"
+	"time"
+
+	"github.com/golang/protobuf/ptypes/timestamp"
 
 	"github.com/TykTechnologies/tyk/config"
+	tyk_analytics "github.com/TykTechnologies/tyk/gateway/analytics_pb"
+	"github.com/golang/protobuf/proto"
+	"gopkg.in/vmihailenco/msgpack.v2"
 )
 
 func TestGeoIPLookup(t *testing.T) {
@@ -182,5 +189,89 @@ func BenchmarkTagHeaders(b *testing.B) {
 		if len(newExistingTags) == 2 {
 			b.Fatal("Existing tags have not been expanded")
 		}
+	}
+}
+
+var fakeRecord = AnalyticsRecord{
+	Method:        http.MethodGet,
+	Host:          "httpbin.org",
+	Path:          "/get",
+	RawPath:       "/get",
+	ContentLength: 0,
+	UserAgent:     "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/32.0.1664.3 Safari/537.36",
+	Day:           7,
+	Month:         6,
+	Year:          1982,
+	Hour:          1,
+	ResponseCode:  200,
+	APIKey:        "myhash",
+	TimeStamp:     time.Now(),
+	APIVersion:    "v1",
+	APIName:       "httpbin",
+	APIID:         "httpbin",
+	OrgID:         "none",
+	OauthID:       "",
+	RequestTime:   3,
+	Latency: Latency{
+		Total:    10,
+		Upstream: 8,
+	},
+	RawRequest:  "",
+	RawResponse: "",
+	IPAddress:   "127.0.0.1",
+	Geo:         GeoData{},
+	Network:     NetworkStats{},
+	Tags:        nil,
+	Alias:       "",
+	TrackPath:   false,
+	ExpireAt:    time.Time{},
+}
+
+var fakeRecordProto = tyk_analytics.AnalyticsRecord{
+	Method:        http.MethodGet,
+	Host:          "httpbin.org",
+	Path:          "/get",
+	RawPath:       "/get",
+	ContentLength: 0,
+	UserAgent:     "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/32.0.1664.3 Safari/537.36",
+	Day:           7,
+	Month:         6,
+	Year:          1982,
+	Hour:          1,
+	ResponseCode:  200,
+	APIKey:        "myhash",
+	TimeStamp:     &timestamp.Timestamp{Seconds: int64(time.Now().Second())},
+	APIVersion:    "v1",
+	APIName:       "httpbin",
+	APIID:         "httpbin",
+	OrgID:         "none",
+	RequestTime:   3,
+	Latency: &tyk_analytics.AnalyticsRecord_Latency{
+		Total:    10,
+		Upstream: 8,
+	},
+	RawRequest:  "",
+	RawResponse: "",
+	IPAddress:   "127.0.0.1",
+	Geo:         &tyk_analytics.AnalyticsRecord_GeoData{},
+	Network:     &tyk_analytics.AnalyticsRecord_NetworkStats{},
+	Tags:        nil,
+	Alias:       "",
+	TrackPath:   false,
+	ExpireAt:    &timestamp.Timestamp{},
+}
+
+func BenchmarkMsgpackMarshal(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		_, _ = msgpack.Marshal(fakeRecord)
+	}
+}
+
+func BenchmarkGogoProtobuf(b *testing.B) {
+	b.ReportAllocs()
+
+	for i := 0; i < b.N; i++ {
+		_, _ = proto.Marshal(&fakeRecordProto)
 	}
 }