… gateway/analytics_pb/analytics.pb.go
… gateway/proto/analytics.pb.go
… go.sum
… vendor/github.com/certifi/gocertifi/gen.go
… vendor/github.com/golang/protobuf/internal/wire/wire.go
… vendor/github.com/golang/protobuf/proto/buffer.go
… vendor/github.com/golang/protobuf/proto/clone.go
… vendor/github.com/golang/protobuf/proto/decode.go
… vendor/github.com/golang/protobuf/proto/defaults.go
… vendor/github.com/golang/protobuf/proto/deprecated.go
… vendor/github.com/golang/protobuf/proto/discard.go
… vendor/github.com/golang/protobuf/proto/encode.go
… vendor/github.com/golang/protobuf/proto/equal.go
… vendor/github.com/golang/protobuf/proto/extensions.go
… vendor/github.com/golang/protobuf/proto/lib.go
… vendor/github.com/golang/protobuf/proto/message_set.go
… vendor/github.com/golang/protobuf/proto/pointer_reflect.go
… vendor/github.com/golang/protobuf/proto/pointer_unsafe.go
… vendor/github.com/golang/protobuf/proto/properties.go
… vendor/github.com/golang/protobuf/proto/proto.go
… vendor/github.com/golang/protobuf/proto/registry.go
… vendor/github.com/golang/protobuf/proto/table_marshal.go
… vendor/github.com/golang/protobuf/proto/table_merge.go
… vendor/github.com/golang/protobuf/proto/table_unmarshal.go
… vendor/github.com/golang/protobuf/proto/text.go
… vendor/github.com/golang/protobuf/proto/text_decode.go
… vendor/github.com/golang/protobuf/proto/text_encode.go
… vendor/github.com/golang/protobuf/proto/text_parser.go
… vendor/github.com/golang/protobuf/proto/wire.go
… vendor/github.com/golang/protobuf/proto/wrappers.go