TODO:

- [ ] Requires pump update & config setting to not break backwards compatibility.
- [ ] Investigate other serialization methods e.g. flatbuffers for example.

```
go test -run=none -bench=BenchmarkAnalytics ./gateway

goos: darwin
goarch: amd64
pkg: github.com/TykTechnologies/tyk/gateway
BenchmarkAnalyticsMsgpackMarshal-8        149674              7416 ns/op            2112 B/op          8 allocs/op
BenchmarkAnalyticsProtoMarshal-8          625963              1799 ns/op             240 B/op          1 allocs/op
PASS
ok      github.com/TykTechnologies/tyk/gateway  3.024s
```

`hey -n 10000 http://localhost:8080/bench/json/valid`

Before:
```
Requests/sec:	1013.9596
Latency distribution:
  10% in 0.0169 secs
  25% in 0.0266 secs
  50% in 0.0376 secs
  75% in 0.0557 secs
  90% in 0.0923 secs
  95% in 0.1216 secs
  99% in 0.1858 secs
```

After:
```
Requests/sec:	1383.0088
Latency distribution:
  10% in 0.0202 secs
  25% in 0.0274 secs
  50% in 0.0352 secs
  75% in 0.0424 secs
  90% in 0.0498 secs
  95% in 0.0563 secs
  99% in 0.0698 secs
```

---

Summary: 

~35% request throughput improvement
and P99 latency drops from 185ms to 69ms.