@tbuchaillot 2021-08-24
All the relevant information was moved to https://tyktech.atlassian.net/browse/TT-3211. I'll close this PR and open a new one when this is fully implemented.